+++ server/test/integration/BraveAlerterConfiguratorTest/getAlertSessionByPhoneNumbersTest.js
@@ -8,12 +8,12 @@ const db = require('../../../db/db')
 const BraveAlerterConfigurator = require('../../../BraveAlerterConfigurator')
 const { sessionDBFactory, buttonDBFactory } = require('../../testingHelpers')
 
-describe('BraveAlerterConfigurator.js integration tests: getAlertSessionByPhoneNumber', () => {
+describe('BraveAlerterConfigurator.js integration tests: getAlertSessionByPhoneNumbers', () => {
   beforeEach(async () => {
     this.chatbotState = CHATBOT_STATE.WAITING_FOR_CATEGORY
     this.sessionIncidentCategory = '2'
     this.sessionToPhoneNumber = '+13335557777'
-    this.installationResponderPhoneNumbers = ['+17775558888']
+    this.installationResponderPhoneNumbers = ['+17775558888', '+18885554444']
     this.installationIncidentCategories = ['Cat1', 'Cat2', 'Cat3']
     this.language = 'de'
     this.respondedByPhoneNumber = '+11114442222'
@@ -48,10 +48,10 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionByPhoneN
     await db.clearTables()
   })
 
-  it('should return an AlertSession with the values from the DB', async () => {
+  it('when given a matching set of phone numbers, should return an AlertSession with the values from the DB', async () => {
     const braveAlerterConfigurator = new BraveAlerterConfigurator()
     const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
-    const alertSession = await braveAlerter.getAlertSessionByPhoneNumber(this.sessionToPhoneNumber)
+    const alertSession = await braveAlerter.getAlertSessionByPhoneNumbers(this.sessionToPhoneNumber, this.installationResponderPhoneNumbers[1])
 
     expect(alertSession).to.eql(
       new AlertSession(
@@ -65,5 +65,21 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionByPhoneN
         this.language,
       ),
     )
+  })
+
+  it('when given a matching devicePhoneNumber but mismatched responderPhoneNumber, should return null', async () => {
+    const braveAlerterConfigurator = new BraveAlerterConfigurator()
+    const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
+    const alertSession = await braveAlerter.getAlertSessionByPhoneNumbers(this.sessionToPhoneNumber, 'not one of the responderPhoneNumbers')
+
+    expect(alertSession).to.be.null
+  })
+
+  it('when given a matching responderPhoneNumber but mismatched devicePhoneNumber, should return null', async () => {
+    const braveAlerterConfigurator = new BraveAlerterConfigurator()
+    const braveAlerter = braveAlerterConfigurator.createBraveAlerter()
+    const alertSession = await braveAlerter.getAlertSessionByPhoneNumbers('not the device phone number', this.installationResponderPhoneNumbers[0])
+
+    expect(alertSession).to.be.null
   })
 })